requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
brotlipy==0.7.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.5
lxml==4.6.3
pycparser==2.20
pyOpenSSL==20.0.1
PySocks==1.7.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
wincertstore==0.2
